Addition of SafeBooruAPI
The commit/push includes the SafeBooruAPI Project, SafeBooruLoader class and the SBImage Class. The function almost exactly the same as the Pixabay API. The Loader has a QueryBuilder that makes the URL to get the XML data, GetXML uses that URL to retrieve and format the XML, and then ParseXMLAndCreateImageList does what it's name implies, properly parsing and sorting all the XML data, which it then takes the important information from to build SBImages and add them to a list it returns at the end.